+++ modules/default/compliments/compliments.js
@@ -1,4 +1,8 @@
 /* global Cron */
+/* global global */
+
+console.log("window name='"+window.name+"'")
+const compliments_test_mode=(window.name=='jsdom')?true:false
 
 Module.register("compliments", {
 	// Module config defaults.
@@ -12,14 +16,17 @@ Module.register("compliments", {
 		},
 		updateInterval: 30000,
 		remoteFile: null,
+		remoteFileRefreshInterval: null,
 		fadeSpeed: 4000,
 		morningStartTime: 3,
 		morningEndTime: 12,
 		afternoonStartTime: 12,
 		afternoonEndTime: 17,
 		random: true,
-		specialDayUnique: false
+		specialDayUnique: false,
 	},
+	urlSuffix:"",
+	compliments_new:null,
 	lastIndexUsed: -1,
 	// Set currentweather from module
 	currentWeatherType: "",
@@ -41,6 +48,14 @@ Module.register("compliments", {
 			const response = await this.loadComplimentFile();
 			this.config.compliments = JSON.parse(response);
 			this.updateDom();
+			if (this.config.remoteFileRefreshInterval !== null) {
+				this.remoteFileRefreshFunc = setInterval(async () => {
+					const response = await this.loadComplimentFile();
+					this.compliments_new = JSON.parse(response);
+				},
+				this.config.remoteFileRefreshInterval
+				)
+			}
 		}
 		let minute_sync_delay = 1;
 		// loop thru all the configured when events
@@ -185,7 +200,14 @@ Module.register("compliments", {
 	async loadComplimentFile () {
 		const isRemote = this.config.remoteFile.indexOf("http://") === 0 || this.config.remoteFile.indexOf("https://") === 0,
 			url = isRemote ? this.config.remoteFile : this.file(this.config.remoteFile);
-		const response = await fetch(url);
+		// because we may be fetching the same url,
+		// we need to force the server to not give us the cached result
+		// create an extra property (ignored by the server handler) just so the url string is different
+		// that will never be the same, using the ms value of date
+		if(this.config.remoteFileRefreshInterval!=null)
+			this.urlSuffix= "?dummy="+Date.now()
+		//
+		const response = await fetch(url+this.urlSuffix);
 		return await response.text();
 	},
 
@@ -236,6 +258,25 @@ Module.register("compliments", {
 			compliment.lastElementChild.remove();
 			wrapper.appendChild(compliment);
 		}
+		// if a new set of compliments was loaded from the refresh task
+		// we do this here to make sure no other function is using the compliments list
+		if(this.compliments_new){
+			// use them
+			if(JSON.stringify(this.config.compliments)!== JSON.stringify(this.compliments_new)){
+				// only reset if the contents changes
+				this.config.compliments = this.compliments_new
+				// reset the index
+				this.lastIndexUsed = -1;
+			}
+		}
+		// in test mode only
+		//if (compliments_test_mode) {
+			// check for (undocumented) remote file2 to change test new load
+			if(this.config.remoteFile2!== null && this.config.remoteFileRefreshInterval!==null){
+	 			 console.log("running in test");
+	 			 this.config.remoteFile=this.config.remoteFile2
+			}
+		//}
 		return wrapper;
 	},